+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ##### 03.24.2024
 
+> Charts addition
 > Usertypes flow completion
 
 ##### 03.23.2024