+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+##### 03.24.2024
+
+> Charts addition
+> Usertypes flow completion
+
 ##### 03.23.2024
 
+> usertypes ui population
+> Deployment URL update
 > Routing Logic Implementation
 > Tutor Dashboard UI development
 > Dashboard UI development